+++ nautilus_trader/config/live.py
@@ -156,8 +156,8 @@ class TradingNodeConfig(NautilusKernelConfig):
         The live risk engine configuration.
     exec_engine : LiveExecEngineConfig, optional
         The live execution engine configuration.
-    persistence : PersistenceConfig, optional
-        The configuration for enabling persistence via feather files.
+    streaming : StreamingConfig, optional
+        The configuration for streaming to feather files.
     data_clients : dict[str, LiveDataClientConfig], optional
         The data client configurations.
     exec_clients : dict[str, LiveExecClientConfig], optional