+++ tests/unit_tests/backtest/test_backtest_engine.py
@@ -52,7 +52,6 @@ from nautilus_trader.model.orderbook.data import OrderBookDelta
 from nautilus_trader.model.orderbook.data import OrderBookDeltas
 from nautilus_trader.model.orderbook.data import OrderBookSnapshot
 from nautilus_trader.persistence.catalog import DataCatalog
-from nautilus_trader.persistence.streaming import StreamingFeatherWriter
 from nautilus_trader.trading.strategy import Strategy
 from tests.test_kit.stubs import MyData
 from tests.test_kit.stubs.component import TestComponentStubs
@@ -170,9 +169,7 @@ class TestBacktestEngine:
         engine.run()
         engine.dispose()
 
-        for writer in engine.kernel.writers:
-            if isinstance(writer, StreamingFeatherWriter):
-                assert all([f.closed for f in writer._files.values()])
+        assert all([f.closed for f in engine.kernel.writer._files.values()])
 
 
 class TestBacktestEngineData: