+++ RELEASES.md
@@ -4,6 +4,7 @@ Released on TBD.
 
 ### Breaking Changes
 - Renamed `StreamingPersistence` to `StreamingFeatherWriter`
+- Renamed `PersistenceConfig` to `StreamingConfig`
 - Renamed `PersistenceConfig.flush_interval` to `flush_interval_ms`
 - Replaced `UUID4.value` property with `UUID4.to_str()` method